+++ CHANGES.md
@@ -3,6 +3,15 @@ a complete change list, only those that may directly interest or affect users.
 
 # 2016-08-27
 
+* GcodeDriver Gcode Configuration UI
+
+	You can now configure all Gcode commands and RegExs via the driver configuration
+	wizard found in Machine Setup -> Driver -> GcodeDriver. The wizard has two tabs:
+	Serial and Gcode. In the Gcode tab you can choose the tool you want to configure
+	and the command for that tool. By choosing the Default tool you configure the
+	default set of commands which are used for fallbacks when tool specific commands
+	are not found.
+
 * GcodeDriver Commands Now In CDATA
 
 	GcodeDriver commands are switched to use CDATA now, instead of escaped XML. This makes
@@ -37,7 +46,7 @@ a complete change list, only those that may directly interest or affect users.
 	
 	Note that the regex contains named groups. The named groups are used to identify the
 	coordinates of each axis you have defined. You should name the groups with the same
-	names in your <axes> section. In the command above the groups / axes are named
+	names in your axes section. In the command above the groups / axes are named
 	x, y, z and rotation.
 
 # 2016-08-08